+++ src/contract/subcontracts/link.ts
@@ -7,6 +7,7 @@ import type { Result } from '../types'
 export class LinkContract extends BaseContract {
   /**
    * This links a profile to another profile with a given link type.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile that is linking to another profile.
    * @param {string} toProfileId - The profile ID of the profile you want to link to.
    * @param {string} linkType - The type of link.
@@ -17,7 +18,7 @@ export class LinkContract extends BaseContract {
     fromProfileId: string,
     toProfileId: string,
     linkType: string,
-  ): Promise<Result<string>> | never {
+  ): Promise<Result<string, true>> | never {
     const tx = await this.contract.linkProfile(
       fromProfileId,
       toProfileId,
@@ -35,6 +36,25 @@ export class LinkContract extends BaseContract {
   }
 
   /**
+   * This gets the linklist id of a {@link linkProfile} transaction.
+   * @category Link
+   * @param txHash {string} - The transaction hash of the transaction you want to get the linklist id of.
+   * @returns The linklist id of the transaction.
+   */
+  @autoSwitchMainnet()
+  async getLinklistIdByTransaction(
+    txHash: string,
+  ): Promise<Result<string>> | never {
+    const receipt = await this.contract.provider.getTransactionReceipt(txHash)
+
+    const parser = this.parseLog<LinkProfileEvent>(receipt.logs, 'linkProfile')
+
+    return {
+      data: parser.args.linklistId.toString(),
+    }
+  }
+
+  /**
    * This creates a profile for an target address and links the fromProfile to it.
    *
    * This should be only called when the target address doesn't have any profile.
@@ -43,6 +63,7 @@ export class LinkContract extends BaseContract {
    * and set the new profile as the primary profile for this address.
    * The new profile's handle will be set to the address of the target address.
    *
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile that is creating the new profile.
    * @param {string} toAddress - The address of the profile you want to link to.
    * @param {string} linkType - The type of link you want to create. This is a string.
@@ -55,10 +76,13 @@ export class LinkContract extends BaseContract {
     linkType: string,
   ):
     | Promise<
-        Result<{
-          toProfileId: string
-          linklistId: string
-        }>
+        Result<
+          {
+            toProfileId: string
+            linklistId: string
+          },
+          true
+        >
       >
     | never {
     const tx = await this.contract.createThenLinkProfile(
@@ -89,6 +113,7 @@ export class LinkContract extends BaseContract {
 
   /**
    * This removes a link from a profile to another profile.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile that is linking to another profile.
    * @param {string} toProfileId - The profile you want to link to.
    * @param {string} linkType - The type of link.
@@ -99,7 +124,7 @@ export class LinkContract extends BaseContract {
     fromProfileId: string,
     toProfileId: string,
     linkType: string,
-  ): Promise<Result<undefined>> | never {
+  ): Promise<Result<undefined, true>> | never {
     const tx = await this.contract.unlinkProfile(
       fromProfileId,
       toProfileId,
@@ -114,6 +139,7 @@ export class LinkContract extends BaseContract {
 
   /**
    * This returns the linked profile ID of a profile with a given link type.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile you want to get the linked profiles from.
    * @param {string} linkType - The type of link you want to get.
    * @returns An array of profile ids that are linked to the profile id passed in.
@@ -129,7 +155,6 @@ export class LinkContract extends BaseContract {
     )
     return {
       data: linkList.map((link) => link.toNumber().toString()),
-      transactionHash: undefined,
     }
   }
 
@@ -137,6 +162,7 @@ export class LinkContract extends BaseContract {
 
   /**
    * This links a profile to an address with a given link type.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile that is linking to the address.
    * @param {string} toAddress - The address of the profile you want to link to.
    * @param {string} linkType - The type of link.
@@ -165,6 +191,7 @@ export class LinkContract extends BaseContract {
 
   /**
    * This links a profile to any URI with a given link type.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile that is linking to another profile.
    * @param {string} toUri - The URI of the profile you want to link to.
    * @param {string} linkType - The type of link you want to create. This is a string that you can define yourself.
@@ -188,6 +215,7 @@ export class LinkContract extends BaseContract {
 
   /**
    * This links a profile to an ERC721 token with a given link type.
+   * @category Link
    * @param {string} fromProfileId - The profile ID of the profile you want to link from.
    * @param {string} toTokenAddress - The address of the ERC721 token you want to link to.
    * @param {string} toTokenId - The token ID of the ERC721 token you want to link to.