+++ src/contract/subcontracts/profile.ts
@@ -8,6 +8,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This creates a new profile for an address, and returns the ID of the newly created profile.
    * When the profile is the first profile created for an address, the address will be set as the primary profile.
+   * @category Profile
    * @param {string} owner - The Ethereum address of the profile owner.
    * @param {string} handle - The handle of the profile you want to create.
    * @param {string} uri - The URI of the file.
@@ -18,7 +19,7 @@ export class ProfileContract extends BaseContract {
     owner: string,
     handle: string,
     uri: string,
-  ): Promise<Result<string>> | never {
+  ): Promise<Result<string, true>> | never {
     this.validateHandleFormat(handle)
 
     const tx = await this.contract.createProfile({
@@ -44,6 +45,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This changes a profile's handle.
+   * @category Profile
    * @param {string} profileId - The profile ID of the user you want to set the handle for.
    * @param {string} handle - The handle you want to set.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
@@ -52,7 +54,7 @@ export class ProfileContract extends BaseContract {
   async setHandle(
     profileId: string,
     handle: string,
-  ): Promise<Result<undefined>> | never {
+  ): Promise<Result<undefined, true>> | never {
     this.validateHandleFormat(handle)
 
     const tx = await this.contract.setHandle(profileId, handle)
@@ -65,6 +67,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This changes a profile's URI.
+   * @category Profile
    * @param profileId - The profile ID of the user you want to set the URI for.
    * @param uri - The URI you want to set.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
@@ -73,7 +76,7 @@ export class ProfileContract extends BaseContract {
   async setProfileUri(
     profileId: string,
     uri: string,
-  ): Promise<Result<undefined>> | never {
+  ): Promise<Result<undefined, true>> | never {
     const tx = await this.contract.setProfileUri(profileId, uri)
     const receipt = await tx.wait()
     return {
@@ -84,6 +87,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This sets the social token for a profile
+   * @category Profile
    * @param {string} profileId - The profile ID of the user you want to set the social token for.
    * @param {string} socialToken - The token address you want to set for the profile.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
@@ -92,7 +96,7 @@ export class ProfileContract extends BaseContract {
   async setSocialToken(
     profileId: string,
     socialToken: string,
-  ): Promise<Result<undefined>> | never {
+  ): Promise<Result<undefined, true>> | never {
     const tx = await this.contract.setSocialToken(profileId, socialToken)
     const receipt = await tx.wait()
     return {
@@ -103,13 +107,14 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This sets the primary profile ID for the user.
+   * @category Profile
    * @param {string} profileId - The profile ID of the profile you want to set as primary.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
    */
   @autoSwitchMainnet()
   async setPrimaryProfileId(
     profileId: string,
-  ): Promise<Result<undefined>> | never {
+  ): Promise<Result<undefined, true>> | never {
     const tx = await this.contract.setPrimaryProfileId(profileId)
     const receipt = await tx.wait()
     return {
@@ -120,6 +125,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns the primary profile ID of the given address.
+   * @category Profile
    * @param {string} address - The address of the user you want to get the primary profile ID for.
    * @returns The profileId of the primary profile of the address.
    */
@@ -133,6 +139,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns a boolean indicating whether the profileId is a primary profileId of an address.
+   * @category Profile
    * @param {string} profileId - The profile ID of the profile you want to check.
    * @returns A boolean value.
    */
@@ -148,6 +155,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns a profile given its handle.
+   * @category Profile
    * @param {string} handle - The handle of the profile you want to get the content of.
    * @returns The profile with the given handle.
    */
@@ -168,6 +176,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns a profile given its profileId.
+   * @category Profile
    * @param {string} profileId - The profile ID of the profile you want to get.
    * @returns The profile with the given profileId.
    */
@@ -187,6 +196,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns the handle of a profile.
+   * @category Profile
    * @param {string} profileId - The profileId of the profile you want to get the handle for.
    * @returns The handle of the profile.
    */
@@ -200,6 +210,7 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This returns the URI of a profile.
+   * @category Profile
    * @param {string} profileId - The profile ID of the profile you want to get the URI for.
    * @returns The URI of the profile.
    */
@@ -212,7 +223,30 @@ export class ProfileContract extends BaseContract {
   }
 
   /**
-   * This validates if a handle is in correct format.
+   * This returns the profile given a {@link createProfile} transaction hash.
+   * @category Profile
+   * @param txHash - The transaction hash of the {@link createProfile} transaction.
+   * @returns The profileId of the profile that was created.
+   */
+  @autoSwitchMainnet()
+  async getProfileByTransaction(
+    txHash: string,
+  ): Promise<Result<Profile>> | never {
+    const receipt = await this.contract.provider.getTransactionReceipt(txHash)
+
+    const parser = this.parseLog<ProfileCreatedEvent>(
+      receipt.logs,
+      'createProfile',
+    )
+
+    const profileId = parser.args.profileId.toString()
+    const result = await this.getProfile(profileId)
+
+    return result
+  }
+
+  /**
+   * This validates if a handle is in a correct format.
    * @param {string} handle - The handle of the profile you want to get the social token for.
    */
   private validateHandleFormat(handle: string): void | never {